feat(stm32mp15-fdts): add STM32M157F boards
STM32MP157F SoC is an update of STM32MP157C, that can run at 800MHz.
In TF-A, this increased frequency is not managed (it is done in OP-TEE)
and both SoCs will start at the nominal frequency: 650MHz.
To create the new boards, just include the STM32MP157C versions, and
update model and compatible. The <board>-fw-config.dts files had already
been sent upstream.

Signed-off-by: Yann Gautier <yann.gautier@st.com>
Change-Id: Ic9fa8f714ba8d8095664c55dbcadfae49e9e2f4f