feat(ti): add message to encrypt tfa during suspend
On some K3 plaforms, BL31 runs from an internal SRAM. During low power
modes like SOC_OFF, IO_ONLY_PLUS_DDR etc, the power to the SRAM
is cut off and only the DDR is active in self-refresh state.

Hence at suspend, BL31 with its context should be encrypted by TIFS and
saved in DDR. Encryption is needed so that the BL31 is not tampered with
while in DRAM.

We only need the encryption function here because the decryption message
will be send by the R5 SPL at resume.

Also introduce the LPM_ENCRYPT_IMAGE[1] capability which signals that FW
has the support to encrypt the image using the TISCI_MSG_LPM_ENCRYPT[2]
tisci message.

The encrypt_tfa message requires that the LPM_SAVE_ADDR[3] message is sent
during boot. This is taken care of by the bootloader.

Link: [1] https://software-dl.ti.com/tisci/esd/latest/2_tisci_msgs/general/core.html#tisci-msg-query-fw-caps
Link: [2] https://software-dl.ti.com/tisci/esd/latest/2_tisci_msgs/pm/lpm.html#lpm-msg-lpm-encrypt
Link: [3] https://software-dl.ti.com/tisci/esd/latest/2_tisci_msgs/pm/lpm.html#lpm-msg-lpm-save-addr

Change-Id: I266472da87dd0821493019b2d9853f8886f33811
Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>
Signed-off-by: Abhash Kumar Jha <a-kumar2@ti.com>