fix(libfdt): fix misra 14.4 and 15.6 violations
Fixed below MISRA violations:

MISRA violation C2012-14.4:
The controlling expression of an if statement and the controlling
expression of an iteration-statement shall have essentially Boolean
type.
Used boolean in control expressions for fixing this violation.

MISRA violation C2012-15.6:
The body of an iteration-statement or a selection-statement
shall be a compound-statement.Enclosed statement body within
the curly braces.

Change-Id: Ic6536e1b749a7aacdf265c632ceea8d344f2e7b5
Signed-off-by: Suraj Kakade <suraj.hanumantkakade@amd.com>