feat(s32g274ardb): load BL31 and BL33 to DDR
Update the platform memory layout to load BL31
and BL33 to DDR instead of SRAM, as it is now
initialized and available.

The `BL31_BASE` and `BL33_BASE` are removed from
`platform_mem_init()` since these images are no
longer loaded into SRAM and should not be initialized
before DDR setup.

Change-Id: Ic12c86b9e66ba710cbd5a42a206f2cd736411956
Signed-off-by: Ghennadi Procopciuc <ghennadi.procopciuc@nxp.com>
Signed-off-by: Andra-Teodora Ilie <andra.ilie@nxp.com>
Signed-off-by: Khristine Andreea Barbulescu <khristineandreea.barbulescu@nxp.com>