feat(s32g274ardb): document DDR integration
Document the integration of DDR driver into
the boot process. Document the relocation
of BL31 and BL33 to DDR instead of SRAM.

Additionally, include the new build parameter
`DDR_FW_BIN_PATH` which must be provided to
specify the path to the DDR firmware binary.
This firmware is used by the DDR driver to
configure and initialize DDR memory.

Change-Id: Ib9fa850926d3dcd745a93eb4aa44846bbdf3e9d3
Signed-off-by: Khristine Andreea Barbulescu <khristineandreea.barbulescu@nxp.com>