feat(spmd): support for extended partition info descriptor
This patch adds the support for extended partition info descriptor
added in FF-A v1.3 ALP2 specification. It enables the LSP, managed by
SPMD, to query the properties of a physical SP.

This is necessary to implement support for live activation of an SP
uniquely identified by its Image UUID.

BREAKING CHANGE: Temporarily fails to build but recovers in subsequent
patches.

Change-Id: Ibb2dab3124ab1838fb954f3d47e11b52e82ca7ca
Signed-off-by: Madhukar Pappireddy <madhukar.pappireddy@arm.com>