refactor(bl1): drop ENABLE_RME check
Starting BL2 in EL3 is not tied to ENABLE_RME but to the boot flow.
In an RME-enabled boot flow, BL2 runs at EL3 and boots BL31. Therefore,
BL2_RUNS_AT_EL3 should be used to handle the transition to the next boot
image.

This change prepares the codebase for later patches that add feature
detection support for RME.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: Ic19b73f1edf5b5ccbd530c662c91e1d38a01fa36