refactor(build): add a default filter list for lib cflags
All libraries use the same set of cflags during compilation,
because the build macro `MAKE_C_LIB` for libraries is hard-coded
to use the common cflags variable `TF_CFLAGS` along with
library-specific `LIB_CFLAGS` in the make command.
However for SIMD crypto extension, some cflags have to
be removed from the build. So add an optional argument
of a list of cflags that needed to be filtered out
for a particular library, to allow more flexibility in the
choice of cflags during compilation.

Change-Id: I53f6f3c6efa28ed4fa2d451270f10f448de2af97
Signed-off-by: Xialin Liu <xialin.liu@arm.com>