fix(rme): do not force set ARM_ARCH_MAJOR/MINOR for ENABLE_RME
ARM_ARCH_MAJOR/MINOR is set to 9.2 when ENABLE_RME=1. It is not clear
why this is required and it is set early in the Makefile before the
platform specific Makefile. Also arch_features are setup based on
ARM_ARCH_MAJOR, ARM_ARCH_MINOR provided by the platform. So there is a
chance this could be overwritten by platform. This change removes
setting ARM_ARCH_MAJOR/MINOR to 9.2. If ENABLE_RME depends on other
architecture features then it must use ENABLE_FEAT option.

Signed-off-by: Arunachalam Ganapathy <arunachalam.ganapathy@arm.com>
Change-Id: I1c74bb53e684d43c62e5f9b7259ebd76add84ff5