feat(versal): add support for TFA_NO_PM configuration
For Versal APU validation in Zephyr CI it is useful to run only one QEMU
instance and not model PLM (Platform Loader and Manager). That's why
implement TFA_NO_PM compilation flag to cover it.  That's why don't call
any PM initialization or communication via IPI with PLM.  Flag is
aligned with implementations done in Versal NET or Versal Gen 2.

Change-Id: I0a4c4c3789f78b649e2961885c87628a599e5536
Signed-off-by: Michal Simek <michal.simek@amd.com>